api/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2023.11.17
click==8.1.7
Deprecated==1.2.14
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-resources==6.1.1
iniconfig==2.0.0
limits==3.7.0
packaging==23.2
passlib==1.7.4
pluggy==1.3.0
psycopg2==2.9.9
pyasn1==0.5.1
pydantic==2.5.3
pydantic_core==2.14.6
pytest==7.4.4
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
slowapi==0.1.8
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
wrapt==1.16.0
